Fix duplicate plugin's basename
I tried to activate a plugin (plugin1) on a site which depend on another plugin (plugin2) who is activated on the network. The plugin2 has been activated on a specific site and then activated network wide.

Whenever I tried to activate plugin1, the activation failed with Plugin dependencies telling me that the requirements was not met. It turn out that during the activation check self::$active_plugins contained the plugin2's basename twice (because of the double activation on a site and network wide).

Removing duplicates with array_unique in self::$active_plugins in check_activation fix the issue.